requirements.txt
requests>=2.28.0
pyyaml>=6.0
numpy>=1.22.0
urllib3>=1.26.0
paramiko>=2.10.0
